Make FinallyScope public
- Storing it as an IDisposable boxes it
