+++ src/QuikGraph.CommonInternals/QuikGraph.CommonInternals.projitems
@@ -12,7 +12,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Compatibility\StringUtils.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Externs\JetBrains.Annotations.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Signing\PublicKey.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)DisposableHelpers.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MathUtils.cs" />
   </ItemGroup>
 </Project>
